requirements.txt
click==8.2.1
colorama==0.4.6
pillow==11.3.0
svgwrite==1.4.3
Tree==0.2.4
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
praw==7.7.1
numpy==1.24.3
python-dotenv==1.0.0